[ ` criticla hotfix 🚑 ` ] : -  fix client repository - separate navbar - fix buttons sign in
