+++ src/app/dashboard/page.tsx
@@ -1,12 +1,4 @@
-'use client'
-
-import { UserRepository } from '@lib/repositories/User.repository'
-import { useEffect } from 'react'
-
 const DashboardPage = (): JSX.Element => {
-  useEffect(() => {
-    UserRepository.getUser().catch(() => {})
-  })
   return (
     <>
       <h1>Dashboard</h1>
