+++ src/layouts/PageWithoutAuth.layout.tsx
@@ -1,4 +1,3 @@
-import { UserServerRepository } from '@lib/repositories/UserServer.repository'
 import { type FC } from 'react'
 import PageLayout, { type IPropsPageLayout } from './Page.layout'
